fix(adapters): resolving correct templateDir
The conditional expression which decides if the passed template name is a path or not resolved the template dir in the wrong order. This fix will change the order. As a result if the template name is a path, the templateDir will be resolved from the template name. Elsewhere it will use the default template dir provided by the Module.forRoot