Fixed handlebars template caching
When using a hierarchical structure for the template folder, in which some files have the same name, only the template that gets rendered first will be used. Any subsequent calls will use the first template instead of the one actually passed, due to the precompiling step using only the file name as the key to the lookup object. This commit changes the key to the full relative path to fix this issue.